+++ libs/superdrops/condensation.hpp
@@ -193,7 +193,7 @@ over for loop: for (size_t kk(0); kk < nsupers; ++kk) {[...]} */
   double totmass_condensed(0.0);                     // cumulative change to liquid mass in parcel volume 'dm'
   Kokkos::parallel_reduce(
       Kokkos::TeamThreadRange(team_member, nsupers),
-      [=, *this](int kk, double &mass_condensed)
+      [=, *this](size_t kk, double &mass_condensed)
       {
         const double deltamass(superdrop_mass_change(supers(kk), temp,
                                                      s_ratio, ffactor));
